feat: Add reports to misnaming tests/suites, executing no tests
Add INFO messages (log level 2) when the framework cannot find the test or suite
required. This is done by modifying two functions to have a return value equal
to the number of tests executed. One last check was added at the end, to verify
if any test was executed. Adapt logging macros to be friendlier to use.

Signed-off-by: mafs <miguelafsilva5@gmail.com>