feat(tf): python bao test framework
Initial version of the Python testing framework. It comprises 1. the test setup parsing (configured through a yaml file), 2. the test setup building, and 3. the results parsing. The framework enables the test of multiple setups on QEMU and allows the definition of different levels of logging (different verbose of test results).

Signed-off-by: Diogo21Costa <diogoandreveigacosta@gmail.com>